Defined folder structure for view components
The folders hierarchy will be separeted by components responsibility. Example: If we have a component that is composed by others, create a folder for him. If this component can be splited in others, all these components have to be inside this folder. If this component just extend components from another folders, just import them.